biblio_scikit-learn_poc/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
fonttools==4.61.1
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
joblib==1.5.3
kiwisolver==1.4.9
logging==0.4.9.6
matplotlib==3.10.8
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pipreqs==0.4.13
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3
yarg==0.1.10
